requirements.txt
requests
flask
speechmatics-python
google-cloud-storage
openai
python-dotenv
gunicorn
mailersend
assemblyai